requirements.txt
# PDF Generation
md2pdf==1.0.1
WeasyPrint>=53.0
markdown2>=2.4.0

# Email Handling
secure-smtplib>=0.1.1

# Environment Variables
python-dotenv>=0.19.0

# Optional Dependencies
Pillow>=8.0.0  # For image handling in PDFs 